+++ app/Http/Controllers/ProductController.php
@@ -22,7 +22,7 @@ class ProductController extends Controller
 
     public function create()
     {
-        $lists = CategoryProduct::all();
+        $lists = CategoryProduct::where('company_id', Auth::user()->company_id)->get();
         return view('dashboard.master-data.product.create', ["lists" => $lists]);
     }
 
@@ -61,8 +61,12 @@ class ProductController extends Controller
     public function edit($id)
     {
         $data = Product::findOrFail($id);
-        $lists = CategoryProduct::all();
-        return view('dashboard.master-data.product.edit', ["data" => $data, "lists" => $lists]);
+        if ($data && $data->company_id == Auth::user()->company_id) { 
+            $lists = CategoryProduct::where('company_id', Auth::user()->company_id)->get();
+            return view('dashboard.master-data.product.edit', ["data" => $data, "lists" => $lists]);
+        } else {
+            return view('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
+        }
     }
 
     public function update(Request $request, $id)
@@ -75,34 +79,41 @@ class ProductController extends Controller
         ]);
 
         $data = Product::findOrFail($id);
-
-        if ($request->file('image')) {
-            if ($request->old_image) {
-                Storage::delete($request->old_image);
+        if ($data && $data->company_id == Auth::user()->company_id) { 
+            if ($request->file('image')) {
+                if ($request->old_image) {
+                    Storage::delete($request->old_image);
+                }
+                $validate['image'] = $request->file('image')->storeAs('images/master-data/product', time() . '.' . $request->image->extension());
             }
-            $validate['image'] = $request->file('image')->storeAs('images/master-data/product', time() . '.' . $request->image->extension());
-        }
-
-        $validate['is_available'] = $request["is_available"] ? 1 : 0;
-
-        $update = $data->update($validate);
-
-        if ($update) {
-            return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to update  product");
+    
+            $validate['is_available'] = $request["is_available"] ? 1 : 0;
+    
+            $update = $data->update($validate);
+    
+            if ($update) {
+                return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to update  product");
+            } else {
+                return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to update  product");
+            }
         } else {
-            return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to update  product");
+            return view('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
         }
     }
 
     public function destroy($id)
     {
         $product = Product::find($id);
-        Storage::delete($product->image);
-        $delete =  Product::destroy($id);
-        if ($delete) {
-            return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to delete  product");
+        if ($product && $product->company_id == Auth::user()->company_id) { 
+            Storage::delete($product->image);
+            $delete =  Product::destroy($id);
+            if ($delete) {
+                return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to delete  product");
+            } else {
+                return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to delete  product");
+            }
         } else {
-            return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to delete  product");
+            return view('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
         }
     }
 }
